@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { EventsService } from '../../service/events.service';
-import { NgFor, NgIf, CommonModule } from '@angular/common';
+import { CommonModule } from '@angular/common';
 import { EventList } from '../../types/eventstype';
 import { Observable, Subscription } from 'rxjs';
 import { Router } from '@angular/router';
@@ -10,7 +10,7 @@ import { SafeUrlService } from '../../service/santizer.service';
 @Component({
   selector: 'app-all-events-listing',
   standalone: true,
-  imports: [NgFor, NgIf, CommonModule, RouterModule],
+  imports: [CommonModule, RouterModule],
   templateUrl: './all-events-listing.component.html',
   styleUrls: ['./all-events-listing.component.css']
 })
